Move DOM and subscribers logic to different modules
